Probably it's easier to review this PR commit-by-commit.

This PR removes three static fields that were growing endlessly:
1. `BaseFaker.PROVIDERS_MAP`
2. `BaseFaker.CLASSES`
3. `FakeValuesService.REGEXP2SUPPLIER_MAP`