@snuyanzin 2024-06-29
Thanks for working on this

I guess we also need to run some measurement tests to see what this change brings to us
I will run it today and post the result later