@asolntsev 2024-06-29
> we also need to run some measurement tests 

At least I run `Issue1178Test`, and it takes ~the same time in `main` branch and `fix/memory-leak`.