@snuyanzin 2024-06-30
there is no significant changes
probably on only one improvement for  `DatafakerSimpleMethods.numberBetween` about 10%, others are more or less same
before 
```
Benchmark                              Mode  Cnt       Score     Error   Units
DatafakerSimpleMethods.firstname      thrpt   10    9383.836 ± 110.342  ops/ms
DatafakerSimpleMethods.fullname       thrpt   10    3108.838 ±  27.933  ops/ms
DatafakerSimpleMethods.numberBetween  thrpt   10  116170.962 ± 705.899  ops/ms
DatafakerSimpleMethods.streetAddress  thrpt   10    2213.100 ±  13.656  ops/ms

Benchmark                        Mode  Cnt      Score      Error   Units
DatafakerExpressions.bothify    thrpt   10  13654.868 ±   78.549  ops/ms
DatafakerExpressions.letterify  thrpt   10  18601.883 ± 1715.311  ops/ms
DatafakerExpressions.numerify   thrpt   10  16980.816 ±  972.231  ops/ms
DatafakerExpressions.regexify   thrpt   10  19546.992 ± 2126.820  ops/ms


Benchmark                  Mode  Cnt    Score    Error   Units
DatafakerCsvjson.csvJson  thrpt   10  460.369 ± 44.570  ops/ms
```
after
```
===============================================================
Benchmark                              Mode  Cnt       Score     Error   Units
DatafakerSimpleMethods.firstname      thrpt   10    9612.673 ± 171.756  ops/ms
DatafakerSimpleMethods.fullname       thrpt   10    3137.267 ±  69.310  ops/ms
DatafakerSimpleMethods.numberBetween  thrpt   10  132792.389 ± 277.564  ops/ms
DatafakerSimpleMethods.streetAddress  thrpt   10    2258.682 ±  31.129  ops/ms

Benchmark                        Mode  Cnt      Score      Error   Units
DatafakerExpressions.bothify    thrpt   10  13748.220 ± 1365.136  ops/ms
DatafakerExpressions.letterify  thrpt   10  20836.751 ±  490.235  ops/ms
DatafakerExpressions.numerify   thrpt   10  19378.905 ± 2497.980  ops/ms
DatafakerExpressions.regexify   thrpt   10  21656.561 ±  271.221  ops/ms

Benchmark                  Mode  Cnt    Score    Error   Units
DatafakerCsvjson.csvJson  thrpt   10  479.511 ± 12.053  ops/ms
```